Fix centraldashboard role bindings
The $(namespace) variable isn't interpolated by kustomize because it's
not defined in params.yaml, and it's not needed because the namespace is
already defined in kustomize.yaml. Without this patch, the
centraldashboard service account does not have any roles or clusterroles
bound because the bindings refer to a service account in a nonexistent
namespace '$(namespace)'.