Remove anonymous envoy filter
Without this patch, users authenticated with dex end up having
",anonymous@kubeflow.org" appended to their usernames, which means that
needs to be appended in their role bindings. This isn't required since
the users are not anonymous, and it's misleading since users aren't
coming from a kubeflow-managed identity provider. This change removes
the filter.